Create api/contestant/team directory, change payment and team schema
