+++ src/app/api/contestant/team/verify/route.ts
@@ -1,14 +1,13 @@
 import { NextRequest, NextResponse } from 'next/server';
-import payOSPaymentService from '@/services/payOSPaymentService';
 import paymentService from '@/services/paymentService';
 
 /**
  * @swagger
- * /api/payment/status-update:
- *   post:
+ * /api/contestant/team/verify:
+ *   patch:
  *     tags:
  *       - Payment
- *     description: Update payment status when the transaction is first created.
+ *     description: Verify and update payment status.
  *     requestBody:
  *       required: true
  *       content:
@@ -16,12 +15,8 @@ import paymentService from '@/services/paymentService';
  *           schema:
  *             type: object
  *             required:
- *               - buyerID
  *               - teamID
  *             properties:
- *               buyerID:
- *                 type: integer
- *                 default: 1
  *               teamID:
  *                 type: integer
  *                 default: 1
@@ -32,19 +27,11 @@ import paymentService from '@/services/paymentService';
  *         description: Error message
  */
 
-export async function POST(req: NextRequest) {
+export async function PATCH(req: NextRequest) {
   try {
     const body = await req.json();
-    body.description = 'OUTRSPACE8';
-    body.items = [
-      {
-        name: 'Phí tham gia Outrspace',
-        quantity: 1,
-        price: 10000,
-      },
-    ];
+    const response = await paymentService.verifyPayment(body);
 
-    const response = await paymentService.createTransaction(body);
     if (response === undefined) throw new Error('Empty response');
     return NextResponse.json(response.responseBody(), { status: response.status });
   } catch (err: any) {
